db migration

+++ src/server/db/item/schema.ts
@@ -13,7 +13,7 @@ export const itemTable = pgTable("item", {
   id: serial("id").primaryKey(),
   name: varchar("name", { length: 256 }).notNull(),
   description: varchar("description", { length: 256 }),
-  categoryId: numeric("category_id")
+  categoryId: integer("category_id")
     .notNull()
     .references(() => categoryTable.id),
   price: numeric("price", { precision: 8, scale: 2 }).notNull(),
